adjust testcase
