--- testcase/hello_arm_cortex_a9/qemu.bat
@@ -1,3 +1,0 @@
-@echo off
-qemu-system-arm --version
-qemu-system-arm -M vexpress-a9 -kernel cortex-a9.elf -serial stdio -m 512
