+++ testcase/hello_arm_cortex_r52/link.ld
@@ -1,40 +1,44 @@
-ENTRY(_start)
+OUTPUT_FORMAT("elf32-littlearm", "elf32-littlearm", "elf32-littlearm")
+OUTPUT_ARCH(arm)
+ENTRY(_Reset)
 
 MEMORY
 {
-    FLASH (rx) : ORIGIN = 0x08000000, LENGTH = 256K
-    RAM (rwx)  : ORIGIN = 0x10000000, LENGTH = 64K
+    ROM (rx) : ORIGIN = 0x08000000, LENGTH = 512k
+    RAM (rwx): ORIGIN = 0x20000000, LENGTH = 512k
 }
 
 SECTIONS
 {
-    .text :
-    {
-        KEEP(*(.isr_vector))
+    .text : {
+        *(.vector_table)
         *(.text*)
         *(.rodata*)
-        _etext = .;
-    } > FLASH
-
-    .data : AT (_etext)
+     } > ROM
+    _text_end = .;
+    .data : AT(ADDR(.text) + SIZEOF(.text))
     {
-        _data = .;
-        *(.data*)
-        _edata = .;
-    } > RAM
-
-    .bss :
-    {
-        _bss = .;
-        *(.bss*)
-        *(COMMON)
-        _ebss = .;
-    } > RAM
-
-    .stack (NOLOAD) :
-    {
+        _data_start = .;
+        *(.data)
         . = ALIGN(8);
-        _stack_base = .;
-        _stack_top = . + 0x1000; /* 4KB stack */
+        _data_end = .;
     } > RAM
+    .bss : {
+        _bss_start = .;
+        *(.bss)
+        . = ALIGN(8);
+        _bss_end = .;
+    } > RAM
+
+    _fiq_stack_start = ADDR(.bss) + SIZEOF(.bss);
+    _fiq_stack_end = _fiq_stack_start + 0x1000; /* 64 KB */
+
+    _irq_stack_start = _fiq_stack_end;
+    _irq_stack_end = _irq_stack_start + 0x1000; /* 64 KB */
+
+    _stack_start = _irq_stack_end;
+    _stack_end = _stack_start + 0x1000; /* 64 KB */
+
+    _irq_stack_size = _irq_stack_end - _irq_stack_start;
+    _fiq_stack_size = _fiq_stack_end - _fiq_stack_start;
 }
